feat: refactor ui architecture
