+++ src/app/not-found.tsx
@@ -1,6 +1,6 @@
 import {Box, Text} from '@chakra-ui/react';
 import Link from 'next/link';
-import {Button} from '@/components/Button';
+import {Button} from '@/components/ui/Button';
 
 export default function NotFound() {
   return (
